testing/web-platform/tests/encoding-detection/tr-windows-1254.tentative.html

<!doctype html>
<title>tr windows-1254</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>Bu bir karakter kodlama testidir. Latince karakterleri kullanan baz� dillerde karar vermek i�in daha fazla girdiye ihtiyac�m�z var.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1254", 'Expected windows-1254');
    }, "Check detection result");
      done();
};
</script>
